# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorDTO:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
  dto.ForgotPasswordDTO:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  dto.LoginDTO:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  dto.MessageDTO:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.RegisterDTO:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.ResetPasswordDTO:
    properties:
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - new_password
    - token
    type: object
...
  dto.UserResponseDTO:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:4200
info:
  contact: {}
  description: REST API для управления коллекционными предметами с JWT авторизацией,
    OAuth 2.0, MinIO и Swagger.
  title: NewYear API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Генерирует одноразовый токен сброса пароля (действует 15 минут).
        В режиме разработки токен сохраняется в папку tmp-mails/. Ответ всегда одинаковый
        — чтобы нельзя было узнать, существует ли email в системе.
      parameters:
      - description: Email пользователя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "400":
          description: Неверный формат email
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Запрос сброса пароля
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Аутентификация по email и паролю. Устанавливает HttpOnly cookies:
        access_token (15 мин) и refresh_token (7 дней).'
      parameters:
      - description: Данные для входа
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Вход в систему
      tags:
      - Auth
  /auth/logout:
    post:
      description: Отзывает текущие access и refresh токены в БД. Очищает HttpOnly
        cookies.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Выход из текущей сессии
      tags:
      - Auth
  /auth/logout-all:
    post:
      description: Отзывает все access и refresh токены пользователя во всех сессиях.
        Очищает HttpOnly cookies.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Выход со всех устройств
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Обновляет access_token и refresh_token используя refresh_token
        из cookie. Старые токены отзываются в БД.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "401":
          description: Refresh token не найден или недействителен
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Обновление пары токенов
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя по email и паролю. Пароль хешируется
        через bcrypt с уникальной солью.
      parameters:
      - description: Данные для регистрации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'example: {\"message\":\"User registered successfully\",\"user\":{\"id\":\"uuid\",\"email\":\"user@example.com\",\"created_at\":\"2024-01-01T00:00:00Z\"}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат данных или пользователь уже существует
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по токену из файла tmp-mails/. Токен
        одноразовый, действует 15 минут. После сброса все сессии пользователя завершаются.
      parameters:
      - description: Токен и новый пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageDTO'
        "400":
          description: Неверный или просроченный токен
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      summary: Установка нового пароля
      tags:
      - Auth
  /auth/whoami:
    get:
      description: Возвращает профиль авторизованного пользователя. Требует валидный
        access_token в HttpOnly cookie. Нужен фронтенду для проверки статуса авторизации,
        так как HttpOnly cookies недоступны из JavaScript.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - CookieAuth: []
      summary: Данные текущего пользователя
      tags:
      - Auth
...
securityDefinitions:
  CookieAuth:
    description: JWT access token через HttpOnly cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"